% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runStudy.R
\name{runStudy}
\alias{runStudy}
\title{Run the P4-C1-021 Alzheimer's disease characterization study}
\usage{
runStudy(cdm, outputFolder, minCellCount = 5)
}
\arguments{
\item{cdm}{A cdm reference object created by `CDMConnector::cdmFromCon`}

\item{outputFolder}{The full path to a folder where the results should be saved}

\item{minCellCount}{The minimum cell count to use in exported results (default 5)}
}
\description{
Run the P4-C1-021 Alzheimer's disease characterization study
}
\examples{
\dontrun{
library(CDMConnector)
con <- DBI::dbConnect(duckdb::duckdb(), eunomiaDir("synpuf-110k"))
cdm <- cdmFromCon(con, "main", "main")
P4C1021::runStudy(cdm, outputFolder = here::here("output"))
}
}
